apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.0
pillow==10.4.0

# Supabase
supabase==2.9.0
gotrue==2.9.1

# Google AI
google-generativeai==0.8.3

# AI/ML (optional - install if using Stable Diffusion)
# diffusers==0.30.0
# torch==2.4.0
# transformers==4.44.0

